requirements.txt
langchain-astradb
sentence-transformers
langchain 
langchain_huggingface
datasets
pypdf
python-dotenv
flask
pandas 

-e .










